requirements.txt
pycrypto
pycryptodome==3.8.0